requirements.txt
gradio_client>=0.8.0
openai
python-dotenv
requests
pydub
datetime
